Improvements

+++ backend/src/modules/user/user.controller.ts
@@ -1,53 +1,47 @@
 import { Hono } from 'hono';
 import { openApi } from 'hono-zod-openapi';
 import z from 'zod';
-import { DeleteUserSchema, UpdateUserSchema, UserSafeSchema } from './user.dto';
+import {
+  DeleteUserSchema,
+  UpdateUserSchema,
+  UserProfileSchema,
+  UserSafeSchema,
+} from './user.dto';
+import {
+  NoContentResponseSchema,
+  NotFoundResponseSchema,
+  ForbiddenResponseSchema,
+} from '@/src/lib/openAPI.types';
 import type { AppContext } from '@/src/types';
 
 const userController = new Hono<AppContext>()
-  .get(
-    '/',
-    openApi({
-      tags: ['User', 'Get All'],
-      responses: { 200: z.array(UserSafeSchema) },
-    }),
-    async (c) => {
-      const userService = c.get('services').user;
-      const users = await userService.getAll();
-      return c.json(
-        users.map((user) => ({
-          id: user.id,
-          name: user.name,
-          email: user.email,
-        })),
-      );
-    },
-  )
   .get(
     '/:id',
     openApi({
       tags: ['User', 'Get By Id'],
       request: {
-        param: z.object({
-          id: z.uuid(),
-        }),
+        param: z.string(), // even though its a UUID we treat it as string so that zod doesnt throw an error when validating the path parameter, and we can return a 404 if the user is not found instead of a 400 bad request error
       },
       responses: {
-        200: UserSafeSchema,
-        404: z.object({ error: z.string() }),
+        200: UserProfileSchema,
+        404: NotFoundResponseSchema,
       },
     }),
     async (c) => {
       const userService = c.get('services').user;
-      const id = c.req.param('id');
-      const user = await userService.getById(id);
+      const id = c.req.valid('param');
+      const user = await userService.getProfileById(id);
       if (!user) {
-        return c.notFound();
+        return c.var.res(404, { message: 'User not found' });
       }
-      return c.json({
+      return c.var.res({
         id: user.id,
         name: user.name,
-        email: user.email,
+        createdAt: user.createdAt,
+        followedByCount: user._count.followedBy,
+        followsCount: user._count.follows,
+        hobbiesCount: user._count.hobbies,
+        hobbySessionsCount: user._count.hobbySessions,
       });
     },
   )
@@ -56,17 +50,28 @@ const userController = new Hono<AppContext>()
     openApi({
       tags: ['User', 'Delete'],
       request: {
-        query: DeleteUserSchema,
+        param: DeleteUserSchema,
       },
       responses: {
-        200: z.object({ deleted: z.boolean() }),
+        204: NoContentResponseSchema,
+        403: ForbiddenResponseSchema,
+        404: NotFoundResponseSchema,
       },
     }),
     async (c) => {
-      const id = c.req.param('id');
+      const id = c.req.valid('param').id;
+      const currentUserId = c.get('userId');
+      if (currentUserId !== id) {
+        return c.var.res(403, { message: 'Unauthorized user' });
+      }
+
       const userService = c.get('services').user;
+
       const deleted = await userService.delete(id);
-      return c.json({ deleted });
+      if (!deleted) {
+        return c.var.res(404, { message: 'User not found' });
+      }
+      return c.var.res(204, null);
     },
   )
   .patch(
@@ -75,24 +80,31 @@ const userController = new Hono<AppContext>()
       tags: ['User', 'Update'],
       request: {
         json: UpdateUserSchema,
+        param: z.string(),
       },
       responses: {
-        204: UserSafeSchema,
-        404: z.object({ error: z.string() }),
+        200: UserSafeSchema,
+        403: ForbiddenResponseSchema,
+        404: NotFoundResponseSchema,
       },
     }),
     async (c) => {
       const body = c.req.valid('json');
-      const id = c.req.param('id');
+      const id = c.req.valid('param');
+
+      const currentUserId = c.get('userId');
+      if (currentUserId !== id) {
+        return c.var.res(403, { message: 'Unauthorized user' });
+      }
+
       const userService = c.get('services').user;
       const updatedUser = await userService.update(id, body);
 
       if (!updatedUser) {
-        return c.notFound();
+        return c.var.res(404, { message: 'User not found' });
       }
 
-      c.status(204);
-      return c.json({
+      return c.var.res({
         id: updatedUser.id,
         name: updatedUser.name,
         email: updatedUser.email,
